+++ tests/aggregate_routing.rs
@@ -55,6 +55,7 @@ fn aggregate_publish_surfaces_three_subs_in_order() {
                 Ok(Some(StreamPacket::Metadata(_))) => {
                     tx.send(Received::Metadata).unwrap();
                 }
+                Ok(Some(StreamPacket::Command(_))) => {}
                 Ok(None) => {
                     tx.send(Received::End).unwrap();
                     break;
@@ -170,6 +171,7 @@ fn aggregate_timestamp_renormalisation_reaches_session() {
                 Ok(Some(StreamPacket::Metadata(_))) => {
                     tx.send(Received::Metadata).unwrap();
                 }
+                Ok(Some(StreamPacket::Command(_))) => {}
                 Ok(None) => {
                     tx.send(Received::End).unwrap();
                     break;
@@ -248,6 +250,7 @@ fn aggregate_command_subs_drive_teardown() {
                 Ok(Some(StreamPacket::Metadata(_))) => {
                     tx.send(Received::Metadata).unwrap();
                 }
+                Ok(Some(StreamPacket::Command(_))) => {}
                 Ok(None) => {
                     tx.send(Received::End).unwrap();
                     break;