+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Other
+
+- typed NetStreamCommand (RTMP 1.0 §4.2 play/play2/pause/seek/receiveAudio/receiveVideo) parser + builder; server surfaces inbound ones as StreamPacket::Command
+
 ## [0.0.6](https://github.com/OxideAV/oxideav-rtmp/compare/v0.0.5...v0.0.6) - 2026-06-15
 
 ### Other